requirements.txt
python-dotenv
openai>=1.0.0
pydantic>=2.0.0
fastapi
uvicorn
sendgrid
